chore: move some methods to observer api
