+++ validity-prover/src/app/observer_api.rs
@@ -1,5 +1,6 @@
 use intmax2_client_sdk::external_api::contract::{
-    liquidity_contract::Deposited, rollup_contract::FullBlockWithMeta,
+    liquidity_contract::Deposited,
+    rollup_contract::{DepositLeafInserted, FullBlockWithMeta},
 };
 use intmax2_interfaces::api::validity_prover::interface::DepositInfo;
 use intmax2_zkp::{
@@ -7,8 +8,9 @@ use intmax2_zkp::{
     ethereum_types::{bytes32::Bytes32, u256::U256, u32limb_trait::U32LimbTrait},
     utils::leafable::Leafable as _,
 };
+use tracing::instrument;
 
-use super::{error::ObserverError, observer::Observer};
+use super::{check_point_store::EventType, error::ObserverError, observer::Observer};
 
 impl Observer {
     pub async fn get_local_last_deposit_id(&self) -> Result<u64, ObserverError> {
@@ -33,6 +35,165 @@ impl Observer {
             .await?;
         let last_block_number = result.and_then(|r| r.max).map(|i| i as u32);
         Ok(last_block_number.unwrap_or(0))
+    }
+
+    #[instrument(skip(self))]
+    pub async fn get_local_next_event_id(
+        &self,
+        event_type: EventType,
+    ) -> Result<u64, ObserverError> {
+        let next_event_id = match event_type {
+            EventType::Deposited => self.get_local_last_deposit_id().await? + 1,
+            EventType::DepositLeafInserted => self
+                .get_local_last_deposit_index()
+                .await?
+                .map(|i| i as u64 + 1)
+                .unwrap_or(0),
+            EventType::BlockPosted => self.get_local_last_block_number().await? as u64 + 1,
+        };
+        Ok(next_event_id)
+    }
+
+    #[instrument(skip(self))]
+    pub async fn get_local_last_eth_block_number(
+        &self,
+        event_type: EventType,
+    ) -> Result<Option<u64>, ObserverError> {
+        let last_eth_block_number = match event_type {
+            EventType::Deposited => {
+                sqlx::query_scalar!(
+                    r#"
+                    SELECT eth_block_number
+                    FROM deposited_events
+                    WHERE deposit_id = (SELECT MAX(deposit_id) FROM deposited_events)
+                    "#
+                )
+                .fetch_optional(&self.pool)
+                .await?
+            }
+            EventType::DepositLeafInserted => {
+                sqlx::query_scalar!(
+                    r#"
+                    SELECT eth_block_number
+                    FROM deposit_leaf_events
+                    WHERE deposit_index = (SELECT MAX(deposit_index) FROM deposit_leaf_events)
+                    "#
+                )
+                .fetch_optional(&self.pool)
+                .await?
+            }
+            EventType::BlockPosted => {
+                sqlx::query_scalar!(
+                    r#"
+                    SELECT eth_block_number
+                    FROM full_blocks
+                    WHERE block_number = (SELECT MAX(block_number) FROM full_blocks)
+                    "#
+                )
+                .fetch_optional(&self.pool)
+                .await?
+            }
+        };
+        // This is a special case for genesis block
+        if last_eth_block_number == Some(0) {
+            return Ok(None);
+        }
+        Ok(last_eth_block_number.map(|i| i as u64))
+    }
+
+    #[instrument(skip(self))]
+    pub async fn get_onchain_next_event_id(
+        &self,
+        event_type: EventType,
+    ) -> Result<u64, ObserverError> {
+        let next_event_id = match event_type {
+            EventType::Deposited => self.liquidity_contract.get_last_deposit_id().await? + 1,
+            EventType::DepositLeafInserted => {
+                self.rollup_contract.get_next_deposit_index().await? as u64
+            }
+            EventType::BlockPosted => {
+                self.rollup_contract.get_latest_block_number().await? as u64 + 1
+            }
+        };
+        Ok(next_event_id)
+    }
+
+    #[instrument(skip(self))]
+    pub async fn is_synced(&self, event_type: EventType) -> Result<bool, ObserverError> {
+        let local_next_event_id = self.get_local_next_event_id(event_type).await?;
+        let onchain_next_event_id = self.get_onchain_next_event_id(event_type).await?;
+        Ok(local_next_event_id >= onchain_next_event_id)
+    }
+
+    // Util function to get deposit_leaf_inserted events between the specified block and the previous block
+    // This is used to generate validity witness for the block
+    #[instrument(skip(self))]
+    pub async fn get_deposits_between_blocks(
+        &self,
+        block_number: u32,
+    ) -> Result<Option<Vec<DepositLeafInserted>>, ObserverError> {
+        if block_number == 0 {
+            return Ok(Some(Vec::new()));
+        }
+        let prev_block_number = block_number - 1;
+        let local_last_block_number = self.get_local_last_block_number().await?;
+        if block_number > local_last_block_number {
+            // blocks are not ready
+            return Ok(None);
+        }
+        let current_block = self
+            .get_full_block_with_meta(block_number)
+            .await?
+            .ok_or(ObserverError::BlockNotFound(block_number))?;
+        let prev_block = self
+            .get_full_block_with_meta(prev_block_number)
+            .await?
+            .ok_or(ObserverError::BlockNotFound(prev_block_number))?;
+        let local_last_eth_block_number = self
+            .get_local_last_eth_block_number(EventType::DepositLeafInserted)
+            .await?;
+        if local_last_eth_block_number.is_none() {
+            let is_synced = self.is_synced(EventType::DepositLeafInserted).await?;
+            if is_synced {
+                // This means no deposit leaf inserted events though we have synced all events
+                return Ok(Some(Vec::new()));
+            } else {
+                //  We have not synced all events yet
+                return Ok(None);
+            }
+        }
+        let local_last_eth_block_number = local_last_eth_block_number.unwrap();
+        if local_last_eth_block_number < current_block.eth_block_number {
+            let is_synced = self.is_synced(EventType::DepositLeafInserted).await?;
+            if !is_synced {
+                return Ok(None);
+            }
+        }
+        let deposits = sqlx::query!(
+            r#"
+            SELECT deposit_index, deposit_hash, eth_block_number, eth_tx_index
+            FROM deposit_leaf_events
+            WHERE (eth_block_number, eth_tx_index) > ($1, $2)
+            AND (eth_block_number, eth_tx_index) <= ($3, $4)
+            ORDER BY deposit_index
+            "#,
+            prev_block.eth_block_number as i64,
+            prev_block.eth_tx_index as i64,
+            current_block.eth_block_number as i64,
+            current_block.eth_tx_index as i64,
+        )
+        .fetch_all(&self.pool)
+        .await?;
+        let events = deposits
+            .into_iter()
+            .map(|d| DepositLeafInserted {
+                deposit_index: d.deposit_index as u32,
+                deposit_hash: Bytes32::from_bytes_be(&d.deposit_hash).unwrap(),
+                eth_block_number: d.eth_block_number as u64,
+                eth_tx_index: d.eth_tx_index as u64,
+            })
+            .collect();
+        Ok(Some(events))
     }
 
     pub async fn get_next_deposit_index(&self) -> Result<u32, ObserverError> {
